finalized support for exporting reported values only
