+++ src/daily-calibration/daily-calibration.service.ts
@@ -30,20 +30,4 @@ export class DailyCalibrationService {
   async export(dailyTestSummaryIds: string[]): Promise<DailyCalibrationDTO[]> {
     return this.dailyCalibrationByTestSumId(dailyTestSummaryIds);
   }
-
-  async removeNonReportedValues(dailyCalibrationData: DailyCalibrationDTO[]) {
-    dailyCalibrationData.forEach(dto => {
-      delete dto.id;
-      delete dto.dailyTestSumId;
-      delete dto.reportingPeriodId;
-      delete dto.calcOnlineOfflineIndicator;
-      delete dto.calcZeroApsIndicator;
-      delete dto.calcUpscaleApsIndicator;
-      delete dto.calcZeroCalibrationError;
-      delete dto.calcUpscaleCalibrationError;
-      delete dto.userId;
-      delete dto.addDate;
-      delete dto.updateDate;
-    })
-  }
 }
