+++ src/daily-emission/daily-emission.service.ts
@@ -1,6 +1,5 @@
 import { Injectable } from '@nestjs/common';
 
-import { DailyEmissionDTO } from 'src/dto/daily-emission.dto';
 import { EmissionsParamsDTO } from '../dto/emissions.params.dto';
 import { DailyFuelService } from '../daily-fuel/daily-fuel.service';
 import { DailyEmissionRepository } from './daily-emission.repository';
@@ -35,22 +34,5 @@ export class DailyEmissionService {
     }
 
     return dailyEmissionData;
-  }
-
-  async removeNonReportedValues(dailyEmissionData: DailyEmissionDTO[]) {
-    const promises = [];
-    dailyEmissionData.forEach(dto => {
-      promises.push(this.dailyFuelService.removeNonReportedValues(dto.dailyFuelData));
-      delete dto.id;
-      delete dto.monitoringLocationId;
-      delete dto.reportingPeriodId;
-      delete dto.calcTotalDailyEmissions;
-      delete dto.calcTotalOpTime;
-      delete dto.userId;
-      delete dto.addDate;
-      delete dto.updateDate;
-    });
-
-    await Promise.all(promises);
   }
 }
