+++ src/daily-fuel/daily-fuel.service.ts
@@ -14,16 +14,4 @@ export class DailyFuelService {
       repository: this.repository,
     });
   }
-
-  async removeNonReportedValues(dailyFuelData: DailyFuelDTO[]) {
-    dailyFuelData.forEach(dto => {
-      delete dto.id;
-      delete dto.dailyEmissionId;
-      delete dto.reportingPeriodId;
-      delete dto.calcFuelCarbonBurned;
-      delete dto.userId;
-      delete dto.addDate;
-      delete dto.updateDate;
-    })
-  }
 }
