+++ src/daily-test-summary/daily-test-summary.service.ts
@@ -50,22 +50,4 @@ export class DailyTestSummaryService {
 
     return summaries;
   }
-
-  async removeNonReportedValues(dailyTestSummaryData: DailyTestSummaryDTO[]) {
-    const promises = [];
-    dailyTestSummaryData.forEach(dto => {
-      promises.push(this.dailyCalibrationService.removeNonReportedValues(dto.dailyCalibrationData));
-      delete dto.id;
-      delete dto.monitoringLocationId;
-      delete dto.reportingPeriodId;
-      delete dto.monitoringSystemRecordId;
-      delete dto.componentRecordId;
-      delete dto.calcTestResultCode;
-      delete dto.userId;
-      delete dto.addDate;
-      delete dto.updateDate;
-    });
-
-    await Promise.all(promises);
-  }
 }
