+++ src/monitor-hourly-value/monitor-hourly-value.service.ts
@@ -16,26 +16,4 @@ export class MonitorHourlyValueService {
     const results = await this.repository.export(hourIds);
     return this.map.many(results);
   }
-
-  async removeNonReportedValues(monitorHourlyValueData: MonitorHourlyValueDTO[]) {
-    monitorHourlyValueData.forEach(dto => {
-      delete dto.id;
-      delete dto.hourId;
-      delete dto.reportingPeriodId;
-      delete dto.monitoringLocationId;
-      delete dto.monitoringSystemRecordId;
-      delete dto.componentRecordId;
-      delete dto.biasAdjustmentFactor;
-      delete dto.calcAdjustedHrlyValue;
-      delete dto.calcLineStatus;
-      delete dto.calcRataStatus;
-      delete dto.calcDaycalStatus;
-      delete dto.calcLeakStatus;
-      delete dto.calcDayintStatus;
-      delete dto.calcF2lStatus;
-      delete dto.userId;
-      delete dto.addDate;
-      delete dto.updateDate;
-    })
-  }
 }
