+++ src/sorbent-trap/sorbent-trap.service.ts
@@ -4,7 +4,6 @@ import { exportSorbentTrapData } from '../sorbent-trap-functions/export-sorbent-
 import { SorbentTrapRepository } from './sorbent-trap.repository';
 import { hasArrayValues } from '../utils/utils';
 import { SamplingTrainService } from '../sampling-train/sampling-train.service';
-import { SorbentTrapDTO } from 'src/dto/sorbent-trap.dto';
 
 @Injectable()
 export class SorbentTrapService {
@@ -35,24 +34,5 @@ export class SorbentTrapService {
     }
 
     return sorbentTrapData;
-  }
-
-  async removeNonReportedValues(sorbentTrapData: SorbentTrapDTO[]) {
-    const promises = [];
-    sorbentTrapData.forEach(dto => {
-      promises.push(this.samplingTrainService.removeNonReportedValues(dto.samplingTrainData));
-      delete dto.id;
-      delete dto.monitoringLocationId;
-      delete dto.reportingPeriodId;
-      delete dto.monitoringSystemRecordId;
-      delete dto.calcPairedTrapAgreement;
-      delete dto.calcModcCode;
-      delete dto.calcHgConcentration;
-      delete dto.userId;
-      delete dto.addDate;
-      delete dto.updateDate;
-    });
-
-    await Promise.all(promises);
   }
 }
