+++ src/summary-value/summary-value.service.ts
@@ -45,18 +45,4 @@ export class SummaryValueService {
       this.repository,
     );
   }
-
-  async removeNonReportedValues(summaryValueData: SummaryValueDTO[]) {
-    summaryValueData.forEach(dto => {
-      delete dto.id;
-      delete dto.monitoringLocationId;
-      delete dto.reportingPeriodId;
-      delete dto.calcCurrentRptPeriodTotal;
-      delete dto.calcOsTotal;
-      delete dto.calcYearTotal;
-      delete dto.userId;
-      delete dto.addDate;
-      delete dto.updateDate;
-    })
-  }
 }
