+++ src/weekly-system-integrity/weekly-system-integrity.service.ts
@@ -25,17 +25,4 @@ export class WeeklySystemIntegrityService {
     }
     return this.map.many(results);
   }
-
-  async removeNonReportedValues(weeklySystemIntegrityData: WeeklySystemIntegrityDTO[]) {
-    weeklySystemIntegrityData.forEach(dto => {
-      delete dto.id;
-      delete dto.weeklyTestSumId;
-      delete dto.reportingPeriodId;
-      delete dto.calcSystemIntegrityError;
-      delete dto.calcApsInd;
-      delete dto.userId;
-      delete dto.addDate;
-      delete dto.updateDate;
-    })
-  }
 }
